Select category button automatically updates witht category name/icon when selected
